snowflake_queries/shared/requirements.txt
# Core dependencies for CloudZero Snowflake telemetry
snowflake-connector-python>=3.0.0
requests>=2.28.0
python-dateutil>=2.8.0
urllib3>=1.26.0

# Optional: Better JSON performance
simplejson>=3.17.0